Separate encoding and translation responsibilities
