+++ prime-router/src/main/kotlin/encoding/FHIR.kt
@@ -1,72 +1,29 @@
 package gov.cdc.prime.router.encoding
 
-import ca.uhn.hl7v2.model.Message
-import ca.uhn.hl7v2.model.v251.segment.MSH
-import io.github.linuxforhealth.fhir.FHIRContext
-import io.github.linuxforhealth.hl7.ConverterOptions
-import io.github.linuxforhealth.hl7.message.HL7MessageEngine
-import io.github.linuxforhealth.hl7.message.HL7MessageModel
-import io.github.linuxforhealth.hl7.resource.ResourceReader
+import ca.uhn.fhir.context.FhirContext
+import ca.uhn.fhir.parser.IParser
 import org.apache.logging.log4j.kotlin.Logging
 import org.hl7.fhir.r4.model.Bundle
 
+/**
+ * Extend the fhir Bundle object with a default encoder
+ */
 fun Bundle.encode(): String {
     return FHIR.encode(this)
 }
 
+/**
+ * A set of behaviors and defaults for FHIR encoding, decoding, and translation.
+ */
 object FHIR : Logging {
+    val defaultContext = FhirContext.forR4()
+    val defaultParser = defaultContext.newJsonParser()
 
-    val defaultMessageTemplates = ResourceReader.getInstance().getMessageTemplates()
-
-    val engine = getMessageEngine()
-
-    fun getMessageEngine(options: ConverterOptions = ConverterOptions.SIMPLE_OPTIONS): HL7MessageEngine {
-        val context = FHIRContext(options.isPrettyPrint(), options.isValidateResource(), options.getProperties())
-
-        return HL7MessageEngine(context, options.getBundleType())
+    fun encode(bundle: Bundle, parser: IParser = defaultParser): String {
+        return parser.encodeResourceToString(bundle)
     }
 
-    fun getMessageType(hl7message: Message): String {
-        val header = hl7message.get("MSH")
-        check(header is MSH)
-        return header.getMessageType().getMsg1_MessageCode().getValue() +
-            "_" +
-            header.getMessageType().getMsg2_TriggerEvent().getValue()
-    }
-
-    fun getHL7MessageModel(
-        messageType: String,
-        messageTemplates: Map<String, HL7MessageModel>,
-    ): HL7MessageModel {
-        return messageTemplates.get(messageType)?.let {
-            it
-        } ?: throw UnsupportedOperationException("Message type not yet supported $messageType")
-    }
-
-    // Custom translations can be handled by adding a customized message type
-    // such as test_ORU_R01 to the list in config.properties. Then creating the accompanying files, such as
-    // hl7/message/test_ORU_R01. Any files in fhir_mapping/hl7/resource will override the defaults allowing customization
-    // of resources like Address or Organization.
-    fun translate(
-        hl7Message: Message,
-        messageType: String = getMessageType(hl7Message),
-        messageTemplates: Map<String, HL7MessageModel> = defaultMessageTemplates,
-        engine: HL7MessageEngine = FHIR.engine,
-    ): Bundle {
-        // NOTE:
-        // extracted from
-        // https://github.com/LinuxForHealth/hl7v2-fhir-converter/blob/d5e43fffa96654e7c5bc896e020ff2fa8aac4ff2/src/main/java/io/github/linuxforhealth/hl7/HL7ToFHIRConverter.java#L135-L159
-        // If timezone specification is needed it can be provided via a custom HL7MessageEngine with a custom FHIRContext that has the time zone ID set
-
-        return getHL7MessageModel(messageType, messageTemplates).convert(hl7Message, engine)
-    }
-
-    fun encode(bundle: Bundle): String {
-        return getMessageEngine().getFHIRContext().encodeResourceToString(bundle)
-    }
-
-    fun decode(json: String): Bundle {
-        val parser = engine.getFHIRContext().parser
+    fun decode(json: String, parser: IParser = defaultParser): Bundle {
         return parser.parseResource(Bundle::class.java, json)
     }
 }
